RONDB-842: Fix test case testNodeRestart -n TransStallTimeout T1
At first this was due to a merge issue. In Ndb the timeout handling was
changed to no longer timeout after sending COMMIT and COMPLETE. However
RonDB had kept this behaviour. So at first this was changed to align
RonDB with Ndb behaviour.

Second there was a bug in both RonDB and Ndb where copyApi needed to
initialise failureNr in the Api connect record.

Thirdly the autotest in RonDB uses a deadlock timeout of 5 seconds
instead of 2.5 seconds, to ensure this doesn't cause issues we extended
the wait in this test case to 8 seconds.

Fourthly a bit more debugging printouts was added and also less frequent
resend of the COMMIT and COMPLETE signals.