RONDB-842: Handle DISK_REORG state better
In the DISK_REORG state there will be an old disk record and a new
larger disk record. This happens when the row grows and can no longer
be stored in its original page and instead a new page is used.

There are two variables referring to the disk page. The first one is
the normal one used, m_disk_callback_page. This is used in all cases
except in the case of the DISK_REORG. In the DISK_REORG one instead
commits the row in page referred to by m_extra_disk_callback_page.

These settings are retrieved in the commit phase as well as in the
prepare phase. These variables refer to the actual memory page used
and they can be different in the commit phase from what they were
in the prepare phase.

In exec_tup_commit we always set m_disk_callback_page even in the
DISK_REORG case. This was not correct. This meant that we got
some lengths from the wrong page and could even crash in rare
cases.

The first call to exec_tup_commit is made from DBLQH and will
always set the pageid to RNIL, in this case we initialise the
value to both the above variables to RNIL. We rely on
retrieve_data_page to set these variables correctly. Finally
we have to get pageid in disk_page_log_buffer_callback from
m_extra_disk_callback_page if != RNIL and otherwise from
m_disk_callback_page. This is handled already correctly in
the case when there is no real-time break while getting UNDO
log space.