RONDB-842: The state DISK_REORG requires release of BUSY state on both pages
DISK_REORG has a new page and an old page and both in the prepare phase and
in the commit phase we need to lock both those pages. If we fail with loading
the extra disk page in the prepare phase, we still need to release the BUSY
state of the old page. The consequence of not releasing the BUSY state is that
LCPs cannot proceed and will hang and eventually crash the data node.

In the commit phase we have a special situation when we first update the row
such that we enter the DISK_REORG state, next we delete the row and finally
after that we commit. In this case we didn't release the BUSY state on the
new page (the extra page). We use the get_page call with DEREF_REQ to release
the busy state also here.