RONDB-842: A potential reason for GCP stop is that multi trp setup hangs
The reason for this hang is that we have sent a delayed signal to send a
new SWITCH_MULTI_TRP_REQ and the node fails while the signal is delayed.
In the case that the node fails in the process we need to avoid sending
SWITCH_MULTI_TRP_REQ and thus start the switch process. Doing so will
have the effect that the multi trp setup isn't done when the node is
ready, since it is already started.